requirements.txt
flask==3.1.*
gunicorn==23.*
requests==2.32.*
python-dotenv==1.1.*
